+++ src/extension/app/plugins/reload/reload.js
@@ -10,9 +10,9 @@
  * governing permissions and limitations under the License.
  */
 
-import { EVENTS, MODALS } from '../../constants.js';
+import { MODALS } from '../../constants.js';
+import { SidekickPlugin } from '../../components/plugin/plugin.js';
 import { newTab } from '../../utils/browser.js';
-import { EventBus } from '../../utils/event-bus.js';
 import { i18n } from '../../utils/i18n.js';
 
 /**
@@ -26,10 +26,10 @@ import { i18n } from '../../utils/i18n.js';
 /**
  * Creates the reload plugin
  * @param {AppStore} appStore The app store
- * @returns {CorePlugin} The reload plugin
+ * @returns {SidekickPlugin} The reload plugin
  */
 export function createReloadPlugin(appStore) {
-  return {
+  return new SidekickPlugin({
     id: 'reload',
     condition: (store) => store.isPreview() || store.isDev(),
     button: {
@@ -46,18 +46,17 @@ export function createReloadPlugin(appStore) {
           appStore.hideWait();
           appStore.reloadPage(newTab(evt));
         } catch (e) {
-          EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-            detail: {
-              type: MODALS.ERROR,
-              data: {
-                message: i18n(appStore.languageDict, 'reload_failure'),
-              },
+          appStore.showModal({
+            type: MODALS.ERROR,
+            data: {
+              message: appStore.i18n('reload_failure'),
             },
-          }));
+          });
         }
       },
       isEnabled: (store) => store.isAuthorized('preview', 'write')
         && store.status.edit && store.status.edit.url, // enable only if edit url exists
     },
-  };
+    appStore,
+  });
 }