+++ src/extension/app/store/app.js
@@ -22,10 +22,13 @@ import {
 } from '../utils/browser.js';
 import { EventBus } from '../utils/event-bus.js';
 import {
-  ENVS, EVENTS, EXTERNAL_EVENTS, MODALS,
+  ENVS, EVENTS, EXTERNAL_EVENTS, MODALS, MODAL_EVENTS,
 } from '../constants.js';
 import { pluginFactory } from '../plugins/plugin-factory.js';
+import { SidekickPlugin } from '../components/plugin/plugin.js';
 import { KeyboardListener } from '../utils/keyboard.js';
+import { ModalContainer } from '../components/modal/modal-container.js';
+import { ToastContainer } from '../components/toast/toast-container.js';
 
 /**
  * The sidekick configuration object type
@@ -49,6 +52,10 @@ import { KeyboardListener } from '../utils/keyboard.js';
 
 /**
  * @typedef {import('@Types').AdminResponse} AdminResponse
+ */
+
+/**
+ * @typedef {import('@Types').Modal} Modal
  */
 
 /**
@@ -105,13 +112,13 @@ export class AppStore {
 
   /**
    * Dictionary of language keys
-   * @type {Object.<string, CorePlugin>}
+   * @type {Object.<string, SidekickPlugin>}
    */
   @observable accessor corePlugins;
 
   /**
    * Dictionary of language keys
-   * @type {Object.<string, CustomPlugin>}
+   * @type {Object.<string, SidekickPlugin>}
    */
   @observable accessor customPlugins;
 
@@ -186,7 +193,7 @@ export class AppStore {
     this.corePlugins = {};
 
     if (this.siteStore.authorized) {
-      const envPlugin = pluginFactory.createEnvPlugin();
+      const envPlugin = pluginFactory.createEnvPlugin(this);
       const previewPlugin = pluginFactory.createPreviewPlugin(this);
       const reloadPlugin = pluginFactory.createReloadPlugin(this);
       const publishPlugin = pluginFactory.createPublishPlugin(this);
@@ -210,10 +217,11 @@ export class AppStore {
       if (plugins && Array.isArray(plugins)) {
         plugins.forEach((cfg, i) => {
           const {
-            id,
-            title,
+            id = `custom-plugin-${i}`,
+            title = id,
             titleI18n,
             url,
+            pinned,
             passConfig,
             passReferrer,
             isPalette,
@@ -254,10 +262,10 @@ export class AppStore {
             // assemble plugin config
           const plugin = {
             custom: true,
-            id: id || `custom-plugin-${i}`,
+            id,
             condition,
             button: {
-              text: (titleI18n && titleI18n[lang]) || title || '',
+              text: (titleI18n && titleI18n[lang]) || title,
               action: () => {
                 if (url) {
                   const target = url.startsWith('/') ? new URL(url, `https://${innerHost}/`) : new URL(url);
@@ -279,7 +287,7 @@ export class AppStore {
                     }));
                   } else {
                     // open url in new window
-                    window.open(target, `hlx-sk-${id || `custom-plugin-${i}`}`);
+                    this.openPage(target.toString(), `hlx-sk-${id || `custom-plugin-${i}`}`);
                   }
                 } else if (eventName) {
                   // fire custom event
@@ -288,17 +296,24 @@ export class AppStore {
               },
               isDropdown: isContainer,
             },
+            pinned,
             container: containerId,
+            appStore: this,
           };
           // check if this overlaps with a core plugin, if so override the condition only
           const corePlugin = this.corePlugins[plugin.id];
           if (corePlugin) {
             // extend default condition
-            const { condition: defaultCondition } = corePlugin;
-            corePlugin.condition = (s) => defaultCondition(s) && condition(s);
+            const { condition: defaultCondition } = corePlugin.config;
+            corePlugin.config.condition = (s) => defaultCondition(s) && condition(s);
           } else {
             // add custom plugin
-            this.customPlugins[plugin.id] = plugin;
+            const customPlugin = new SidekickPlugin(plugin);
+            if (plugin.container) {
+              this.customPlugins[plugin.container]?.append(customPlugin);
+            } else {
+              this.customPlugins[plugin.id] = customPlugin;
+            }
           }
         });
       }
@@ -497,6 +512,42 @@ export class AppStore {
   }
 
   /**
+   * Creates a modal and appends it to the sidekick.
+   * If a modal is already open, it will be replaced.
+   * @param {Modal} modal The modal to display
+   */
+  showModal(modal) {
+    const existingModal = this.sidekick?.shadowRoot?.querySelector('theme-wrapper').querySelector('modal-container');
+    if (existingModal) {
+      existingModal.remove();
+    }
+
+    const modalContainer = new ModalContainer(modal);
+    this.sidekick?.shadowRoot?.querySelector('theme-wrapper').appendChild(modalContainer);
+
+    return modalContainer;
+  }
+
+  /*
+   * Returns the currebnt environment
+   * @returns {string} the current environment
+   */
+  getEnv() {
+    return [
+      'isEditor',
+      'isPreview',
+      'isLive',
+      'isProd',
+      'isAdmin',
+      'isDev',
+    ]
+      .filter((method) => this[method]())
+      .map((method) => method.substring(2)) // cut off 'is'
+      .join('')
+      .toLowerCase();
+  }
+
+  /**
    * Displays a wait modal
    * @param {string} [message] The message to display
    */
@@ -504,19 +555,18 @@ export class AppStore {
     if (!message) {
       message = this.i18n('please_wait');
     }
-    EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-      detail: {
-        type: MODALS.WAIT,
-        data: { message },
-      },
-    }));
+
+    return this.showModal({
+      type: MODALS.WAIT,
+      data: { message },
+    });
   }
 
   /**
    * Hides the modal
    */
   hideWait() {
-    EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.CLOSE_MODAL));
+    EventBus.instance.dispatchEvent(new CustomEvent(MODAL_EVENTS.CLOSE));
   }
 
   /**
@@ -541,6 +591,7 @@ export class AppStore {
 
   /**
    * Reloads the current page. Abstracted for testing.
+   * @param {boolean} [newTab] Open current page in a new tab
    */
   reloadPage(newTab) {
     if (newTab) {
@@ -615,14 +666,20 @@ export class AppStore {
    * @param {string} [variant] The variant of the toast (optional)
    * @param {number} [timeout] The timeout in milliseconds (optional)
    */
-  showToast(message, variant = 'info', timeout = 2000) {
-    EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.SHOW_TOAST, {
-      detail: {
-        message,
-        variant,
-        timeout,
-      },
-    }));
+  showToast(message, variant = 'info', timeout = 6000) {
+    const existingToast = this.sidekick?.shadowRoot?.querySelector('theme-wrapper').querySelector('toast-container');
+    if (existingToast) {
+      existingToast.remove();
+    }
+
+    const toastContainer = new ToastContainer({
+      message,
+      variant,
+      timeout,
+    });
+    this.sidekick?.shadowRoot?.querySelector('theme-wrapper').appendChild(toastContainer);
+
+    return toastContainer;
   }
 
   /**
@@ -818,27 +875,21 @@ export class AppStore {
         }, { once: true });
         this.fetchStatus();
       } else if (status.webPath.startsWith('/.helix/') && resp.error) {
-        // show detail message only in config update mode
-        EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-          detail: {
-            type: MODALS.ERROR,
-            data: {
-              message: `${this.i18n('error_config_failure')}${resp.error}`,
-            },
+        this.showModal({
+          type: MODALS.ERROR,
+          data: {
+            message: `${this.i18n('error_config_failure')}${resp.error}`,
           },
-        }));
+        });
       } else {
         // eslint-disable-next-line no-console
         console.error(resp);
-
-        EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-          detail: {
-            type: MODALS.ERROR,
-            data: {
-              message: this.i18n('error_preview_failure'),
-            },
+        this.showModal({
+          type: MODALS.ERROR,
+          data: {
+            message: this.i18n('error_preview_failure'),
           },
-        }));
+        });
       }
       return;
     }
@@ -1203,12 +1254,12 @@ export class AppStore {
         }
         if (attempts >= 5) {
           // give up after 5 attempts
-          EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-            detail: {
-              type: MODALS.ERROR,
-              data: { message: this.i18n('error_login_timeout') },
+          this.showModal({
+            type: MODALS.ERROR,
+            data: {
+              message: this.i18n('error_login_timeout'),
             },
-          }));
+          });
           return;
         }
       }
@@ -1251,13 +1302,12 @@ export class AppStore {
           return;
         }
         if (attempts >= 5) {
-          // give up after 5 attempts
-          EventBus.instance.dispatchEvent(new CustomEvent(EVENTS.OPEN_MODAL, {
-            detail: {
-              type: MODALS.ERROR,
-              data: { message: this.i18n('error_logout_error') },
+          this.showModal({
+            type: MODALS.ERROR,
+            data: {
+              message: this.i18n('error_logout_error'),
             },
-          }));
+          });
           return;
         }
       }