+++ src/extension/app/store/site.js
@@ -10,6 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
+// import { property } from 'lit/decorators.js';
+import { observable, action } from 'mobx';
 import { log } from '../../log.js';
 import { getAdminUrl, getAdminFetchOptions } from '../utils/helix-admin.js';
 import { getLanguage } from '../utils/i18n.js';
@@ -169,10 +171,24 @@ export class SiteStore {
   authTokenExpiry;
 
   /**
+   * Has the store been initialized?
+   * @type {boolean}
+   */
+  @observable accessor ready = false;
+
+  /**
    * @param {AppStore} appStore
    */
   constructor(appStore) {
     this.appStore = appStore;
+  }
+
+  /**
+   * Set as initialized.
+   */
+  @action
+  setReady() {
+    this.ready = true;
   }
 
   /**
@@ -275,6 +291,7 @@ export class SiteStore {
     this.project = project;
     this.devUrl = devUrl;
     this.lang = lang || getLanguage();
+    this.setReady();
   }
 
   /**
@@ -288,6 +305,7 @@ export class SiteStore {
       ref: this.ref,
       giturl: this.giturl,
       devUrl: this.devUrl.href,
+      // @ts-ignore
       mountpoint: this.mountpoint,
       mountpoints: this.mountpoints,
       project: this.project,